added winston and performed project restructure

+++ contollers/user_controller.js
@@ -4,6 +4,7 @@ module.exports = {
     getUser: async (req,res)=>{
         try {
             const user = await User.findById(req.user.id).select('-password');
+            if (!user)  return res.status(404).json({ message: "User not found" });
             res.status(200).json(user)
         } catch (error) {
             console.log(error);
@@ -13,7 +14,9 @@ module.exports = {
 
     deleteUser:  async (req,res)=>{
         try {
-            const user = await User.findByIdAndDelete(req.user.id);
+            const user = await User.findById(req.user.id);
+            if (!user)  return res.status(404).json({ message: "User not found" });
+            await User.findByIdAndDelete(req.user.id);
             res.status(200).json("User deleted")
         } catch (error) {
             console.log(error);
